Added convertNumberToString function
Added function to convert a number to a String so that all of the
subclasses could use. The idea is to call this function with guard
instead of using an if-else statement and nesting the other code that
depends on the String.

There are also two constants defined with this change that are used for
determining the number of digits after the decimal place when
formatting a number as a String.  One is used for alpha values and the
other for colour fractions (colours who have values represented between
0 and 1).