Making regular expressions easier to maintain.
Breaking out common parts of regular expressions and storing them into
variables so that they only need to be changed in one place. I’ve also
found that there were some inconsistencies in how some things were
searched for in between different seekers so this will resolve that.

All basic search elements, such as how to find a float between 0 and 1,
will be defined in Seeker. This is then used in all of the sub-classes
in a consistent manner. If it needs to be changed it’s done in one
place and everything else sees that change right away.