Fix to allow plug-in to work on pre 7.3 Xcode
I had made changes to get ready for Swift 3 by changing deprecated
methods but the new way to handle selectors isn’t supported on Xcode
7.2.1 and below so the plug-in wasn’t being compiled. I’ve reverted the
code back and it should now work.

I need to find a way to have it optionally compile the old method or
the new method or else once the new version of Xcode comes out that has
Swift 3 then the people that have 7.2.1 and below won’t have the
plug-in.